add deck to game
